Hex Grid and Mouse
Can make a hex grid of arbitrary hex size and grid radius
Can just make hexes, it doesn't find neighbours or convert a world pos to
a hex.
And have to make them in solid 'circles'. Can't make a line of hexes or a
ring for example.

When you move your mouse over the grid, one hex is coloured silver