Larger Map, Gold
Made the hex grid bigger
Fixed Camera
Changed scaling so it should be consistent when I go fullscreen.

Tower
Split spawn tower into 2 functions.
Can send an event to spawn it.

Sprite Layers
Changed some layers so the right stuff is on top.
Background is 0.0
On top of that is in the positive direction

Gold
When you place a tower, place a gold spawner there too, It should be on
the area around the tower, but it's like this for testing.
It spawns gold every x seconds (3 for testing, 10 in the game)
Gold follows your mouse
Gold Piles
Suck up gold that runs into it.
This will be how you buy stuff.