Preview Towers, Neighbour Spawners
Neighbour Spawners
Gold Spawners are placed around the tower instead of on top of it.
Can't place spawners on towers.
If there is a spawner where you want to place a tower, it is destroyed.

Modification
This is how I was going to tell the spawner to despawn, but I just used
the TowerBuiltEvent instead. Might come back to this

Gold Piles
Have a cap
Send an event when full

Preview Towers
Place where you want to build a tower.
Then you need to drag gold onto it to pay for it.
This is a TowerPreview + GoldPile with the cap set to the cost. When full,
it sends the event and it gets built (just prints for now)

Preview Sprite
Set as a child. So the hex can have its sprite and a separate sprite for
the preview/tower

Preview -> Tower
New system

Tower
Want to change this a bit
Tower isn't attached to the hex. I think it would be better if it were.

Input
Click to make a preview
Space to make a tower
Need to make towers to make spawners to pay for previews.
Once I can break gold piles, I can start with a gold pile and break it to
pay for preview to build towers to make spawners.