Gold Piles, Paying for Towers
Press G to spawn a pile
Press X to remove it

When you remove the pile, its gold gets dumped out
Piles now get attached to the hex.
A pile gets spawned at the center with some starting cash

Gold
Spawning is separate from the GoldSpawner generating it.
So you can spawn gold when a pile breaks.

Towers
When you pay for the tower, it loses its gold pile and tower preview
components and adds a Tower component.
Also the child sprite changes to red.

Bug
If you press x over a tower preview, the pile will be removed, but the
tower preview won't