Shooting
Towers can shoot
They have unlimited range
They pick the closest enemy and shoot straight at them.
An enemy doesn't have to be within range for the tower to shoot at it.

Bullet
Only travel for 1s. That's the limitation on their range.

Enemies
Have 1hp.
When they die, they drop the gold they're carrying.
I think I did this in a dumb way. Setting it as a child and removing the
Gold component so it doesn't interact with the mouse, etc. and then when
it dies, removing the child and re-adding the gold was a pain.
Deleting the gold and spawning a new one probably would've been better.
Although I did learn a bit about 'entity does not exist'

Enemies can only carry 1 gold now.
They used to be able to grab as many as they collided with

Dead
Add to an enemy to kill it next frame
I think I have the ordering right so it gets deleted and then nothing
depends on it.